+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added `csm-utils.cached_property`
+
 ## [0.1.0] - 2026-05-18
 
 Initial version.